[mix_task] Split commands
